@@ -3,65 +3,68 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
+    "start": "react-native start",
     "test": "jest",
-    "postinstall": "cp src/attachWebsocketServer.js node_modules/react-native/local-cli/server/util/"
+    "lint": "eslint ."
   },
   "dependencies": {
+    "amazon-cognito-identity-js": "^3.0.15",
     "amplitude-js": "^4.7.0-react-native",
-    "aws-amplify": "^1.1.26",
-    "aws-amplify-react-native": "^2.1.10",
-    "axios": "^0.18.0",
-    "lodash": "^4.17.11",
+    "aws-amplify": "^1.1.40",
+    "aws-amplify-react-native": "^2.1.19",
+    "axios": "^0.19.0",
+    "install": "^0.13.0",
+    "lodash": "^4.17.15",
     "moment": "^2.24.0",
-    "native-base": "^2.11.0",
-    "react": "16.6.3",
-    "react-native": "0.58.4",
-    "react-native-camera": "^2.11.2",
-    "react-native-dialog": "^5.5.0",
-    "react-native-fingerprint-scanner": "2.6.2",
-    "react-native-gesture-handler": "^1.0.15",
-    "react-native-image-picker": "0.28.1",
+    "native-base": "^2.13.8",
+    "npm": "^6.11.3",
+    "or": "^0.2.0",
+    "react": "16.8.6",
+    "react-native": "0.60.5",
+    "react-native-camera": "^3.4.0",
+    "react-native-dialog": "^5.6.0",
+    "react-native-fingerprint-scanner": "^3.0.0",
+    "react-native-gesture-handler": "^1.4.1",
+    "react-native-image-picker": "^1.1.0",
     "react-native-intercom": "^13.1.0",
-    "react-native-linear-gradient": "^2.5.4",
-    "react-native-maps": "^0.23.0",
-    "react-native-modal": "^9.0.0",
+    "react-native-linear-gradient": "^2.5.6",
+    "react-native-maps": "0.25.0",
+    "react-native-modal": "^11.4.0",
     "react-native-orientation-locker": "^1.1.6",
+    "react-native-reanimated": "^1.2.0",
     "react-native-render-html": "^4.1.2",
     "react-native-smooth-pincode-input": "^1.0.7",
-    "react-native-svg": "^9.8.4",
-    "react-native-video": "^4.3.1",
-    "react-navigation": "^3.2.1",
-    "react-navigation-redux-helpers": "^3.0.0",
-    "react-redux": "^6.0.0",
-    "redux": "^4.0.1",
-    "redux-persist": "^5.10.0",
+    "react-native-svg": "^9.9.4",
+    "react-native-video": "^5.0.2",
+    "react-navigation": "^4.0.7",
+    "react-navigation-redux-helpers": "^3.0.3",
+    "react-navigation-stack": "^1.8.1",
+    "react-navigation-tabs": "^2.5.5",
+    "react-redux": "^7.1.1",
+    "redux": "^4.0.4",
+    "redux-persist": "^6.0.0",
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "24.1.0",
-    "eslint": "^5.13.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-node": "^8.0.1",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-react": "^7.12.4",
-    "eslint-plugin-react-native": "^3.6.0",
-    "jest": "24.1.0",
-    "metro-react-native-babel-preset": "0.51.1",
-    "react-test-renderer": "16.6.3",
-    "reactotron-react-native": "^2.1.5",
-    "reactotron-redux": "^2.1.3"
+    "@babel/core": "^7.6.0",
+    "@babel/runtime": "^7.6.0",
+    "@react-native-community/eslint-config": "^0.0.5",
+    "babel-jest": "^24.9.0",
+    "eslint": "^6.4.0",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-node": "^10.0.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.14.3",
+    "eslint-plugin-react-native": "^3.7.0",
+    "jest": "^24.9.0",
+    "metro-react-native-babel-preset": "^0.56.0",
+    "react-test-renderer": "16.8.6",
+    "reactotron-react-native": "^4.0.0",
+    "reactotron-redux": "^3.1.2"
   },
   "jest": {
     "preset": "react-native"
-  },
-  "rnpm": {
-    "assets": [
-      "./src/assets/fonts/"
-    ]
   }
 }
