@@ -1,4 +1,4 @@
-import React, { PureComponent } from 'react';
+import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Image, TouchableOpacity } from 'react-native';
@@ -14,35 +14,28 @@ import idCheckUtils from 'src/Common/idCheck';
 import EditIcon from 'src/assets/images/Edit.png';
 
 import { userSelector } from 'src/Redux/AppContent';
+import { localAuthSelector } from 'src/Redux/Auth';
+
 
 import { sg } from 'src/Styles';
 import styles from './styles';
 
-class IdCheck extends PureComponent {
-    // const { user } = this.props;
-    state = {};
-
-    componentDidMount() {
-      const { screenProps } = this.props;
-      console.log('ID CHECK SCREEN');
+class IdCheck extends Component {
+  onPressEditItem(type) {
+    const { screenProps } = this.props;
+    console.log(`type: ${type}`);
+    if (type === 'Drivers Licence') {
+      screenProps.navigateTo(routeNames.ID_CHECK_DRIVERS_LICENCE);
+    } else {
+      screenProps.navigateTo(routeNames.ID_CHECK_AUSTRALIAN_PASSPORT);
     }
+  }
 
-
-    onPressEditItem(type) {
-      const { screenProps } = this.props;
-      console.log(`type: ${type}`);
-      if (type === 'Drivers Licence') {
-        screenProps.navigateTo(routeNames.ID_CHECK_DRIVERS_LICENCE);
-      } else {
-        screenProps.navigateTo(routeNames.ID_CHECK_AUSTRALIAN_PASSPORT);
-      }
-    }
-
-    renderButtons() {
-      const { user, screenProps } = this.props;
-      return (
-        <View>
-          {(!user.idCheck.driversLicence || user.idCheck.driversLicence === 'notAttempted')
+  renderButtons() {
+    const { user, screenProps } = this.props;
+    return (
+      <View>
+        {(!user.idCheck.driversLicence || user.idCheck.driversLicence === 'notAttempted')
             && (
             <Button
               block
@@ -58,7 +51,7 @@ class IdCheck extends PureComponent {
             </Button>
             )
           }
-          {(!user.idCheck.australianPassport || user.idCheck.australianPassport === 'notAttempted')
+        {(!user.idCheck.australianPassport || user.idCheck.australianPassport === 'notAttempted')
             && (
             <Button
               block
@@ -74,7 +67,7 @@ class IdCheck extends PureComponent {
             </Button>
             )
           }
-          {/* NEED TO IMPROVE MEDICARE CARD SCREEN AND TEST BEFORE WE TURN THIS ON
+        {/* NEED TO IMPROVE MEDICARE CARD SCREEN AND TEST BEFORE WE TURN THIS ON
 
             (!user.idCheck.medicareCard || user.idCheck.medicareCard === "notAttempted")
           && (user.idCheck.australianPassport === "matched" || user.idCheck.driversLicence === "matched")
@@ -91,31 +84,31 @@ class IdCheck extends PureComponent {
               </Text>
             </Button>
           */}
-        </View>
-      );
-    }
+      </View>
+    );
+  }
 
-    renderItem(docType, status, user) {
-      const matchFailed = status === 'matchFailed';
-      const isVerified = status === 'matched';
-      const type = idCheckUtils.getTypeName(docType);
-      const passportName = user.idCheck.passportMiddleNames
-        ? `${user.idCheck.passportFirstName} ${user.idCheck.passportMiddleNames} ${user.idCheck.passportLastName}`
-        : `${user.idCheck.passportFirstName} ${user.idCheck.passportLastName}`;
-      const dlName = user.idCheck.driversLicenceMiddleNames
-        ? `${user.idCheck.driversLicenceFirstName} ${user.idCheck.driversLicenceMiddleNames} ${user.idCheck.driversLicenceLastName}`
-        : `${user.idCheck.driversLicenceFirstName} ${user.idCheck.driversLicenceLastName}`;
+  renderItem(docType, status, user) {
+    const matchFailed = status === 'matchFailed';
+    const isVerified = status === 'matched';
+    const type = idCheckUtils.getTypeName(docType);
+    const passportName = user.idCheck.passportMiddleNames
+      ? `${user.idCheck.passportFirstName} ${user.idCheck.passportMiddleNames} ${user.idCheck.passportLastName}`
+      : `${user.idCheck.passportFirstName} ${user.idCheck.passportLastName}`;
+    const dlName = user.idCheck.driversLicenceMiddleNames
+      ? `${user.idCheck.driversLicenceFirstName} ${user.idCheck.driversLicenceMiddleNames} ${user.idCheck.driversLicenceLastName}`
+      : `${user.idCheck.driversLicenceFirstName} ${user.idCheck.driversLicenceLastName}`;
 
-      if (matchFailed || isVerified) {
-        return (
-          <View>
-            <Grid style={styles.itemBl}>
-              <Row>
-                <Col>
-                  <Text style={[sg.textBold, sg.fS20, sg.colorDark2, sg.mB10]}>{type}</Text>
+    if (matchFailed || isVerified) {
+      return (
+        <View>
+          <Grid style={styles.itemBl}>
+            <Row>
+              <Col>
+                <Text style={[sg.textBold, sg.fS20, sg.colorDark2, sg.mB10]}>{type}</Text>
 
 
-                  {matchFailed && docType === 'DriversLicence' && (
+                {matchFailed && docType === 'DriversLicence' && (
                   <View>
                     <Text style={[sg.colorDark3]}>{`${user.idCheck.driversLicenceFirstName} ${user.idCheck.driversLicenceLastName}`}</Text>
                     <Text style={[sg.colorDark3, sg.mV5]}>
@@ -127,10 +120,10 @@ class IdCheck extends PureComponent {
                       {user.idCheck.driversLicenceState}
                     </Text>
                   </View>
-                  )
+                )
                   }
 
-                  {matchFailed && docType === 'Passport' && (
+                {matchFailed && docType === 'Passport' && (
                   <View>
                     <Text style={[sg.colorDark3]}>{passportName}</Text>
                     <Text style={[sg.colorDark3, sg.mV5]}>
@@ -138,49 +131,49 @@ class IdCheck extends PureComponent {
                       {user.idCheck.passportNumber}
                     </Text>
                   </View>
-                  )
+                )
                   }
 
-                </Col>
-                <Col style={sg.width20}>
-                  {matchFailed && (
+              </Col>
+              <Col style={sg.width20}>
+                {matchFailed && (
                   <TouchableOpacity onPress={() => this.onPressEditItem(type)}>
                     <Image source={EditIcon} />
                   </TouchableOpacity>
-                  )}
-                </Col>
+                )}
+              </Col>
+            </Row>
+            {isVerified ? (
+              <Row style={[sg.mT15, sg.aICenter]}>
+                <Icon name="ios-close-circle" style={styles.itemStatusIconOk} />
+                <Text style={styles.itemStatusTextOk}>Verified</Text>
               </Row>
-              {isVerified ? (
-                <Row style={[sg.mT15, sg.aICenter]}>
-                  <Icon name="ios-close-circle" style={styles.itemStatusIconOk} />
-                  <Text style={styles.itemStatusTextOk}>Verified</Text>
-                </Row>
-              ) : (
-                <Row style={[sg.mT15, sg.aICenter]}>
-                  <Icon name="ios-close-circle" style={styles.itemStatusIconErr} />
-                  <Text style={styles.itemStatusTextErr}>Details didn’t match</Text>
-                </Row>
-              )}
-            </Grid>
-          </View>
-        );
-      }
-      return null;
+            ) : (
+              <Row style={[sg.mT15, sg.aICenter]}>
+                <Icon name="ios-close-circle" style={styles.itemStatusIconErr} />
+                <Text style={styles.itemStatusTextErr}>Details didn’t match</Text>
+              </Row>
+            )}
+          </Grid>
+        </View>
+      );
     }
+    return null;
+  }
 
-    render() {
-      const { user, screenProps } = this.props;
-      // const { list } = this.state;
-      return (
-        <Content padder contentContainerStyle={[sg.flexGrow, sg.pT0]}>
+  render() {
+    const { user, screenProps } = this.props;
+    // const { list } = this.state;
+    return (
+      <Content padder contentContainerStyle={[sg.flexGrow, sg.pT0]}>
 
+        <View>
+          <Text style={sg.formHeading}>Your ID check</Text>
+        </View>
+        <View style={[sg.spaceBetween]}>
           <View>
-            <Text style={sg.formHeading}>Your ID check</Text>
-          </View>
-          <View style={[sg.spaceBetween]}>
-            <View>
 
-              {(!user.idCheck || user.idCheck.australianPassport !== 'matchFailed' && user.idCheck.driversLicence !== 'matchFailed')
+            {(!user.idCheck || user.idCheck.australianPassport !== 'matchFailed' && user.idCheck.driversLicence !== 'matchFailed')
               && (
               <View>
                 <Text style={sg.formHeadingDescription}>
@@ -189,7 +182,7 @@ class IdCheck extends PureComponent {
               </View>
               )}
 
-              {user.idCheck
+            {user.idCheck
                 && (user.idCheck.australianPassport === 'matchFailed' || user.idCheck.driversLicence === 'matchFailed')
                 && (
                 <View>
@@ -214,33 +207,33 @@ class IdCheck extends PureComponent {
                 */}
                 </View>
                 )}
-            </View>
-            <View>
-              {user.idCheck
+          </View>
+          <View>
+            {user.idCheck
               && user.idCheck.driversLicence !== 'notAttempted'
               && this.renderItem(
                 idCheckUtils.ID_TYPE.DRIVERS_LICENSE,
                 user.idCheck.driversLicence,
                 user,
               )}
-              {user.idCheck
+            {user.idCheck
               && user.idCheck.australianPassport !== 'notAttempted'
               && this.renderItem(
                 idCheckUtils.ID_TYPE.PASSPORT,
                 user.idCheck.australianPassport,
                 user,
               )}
-              {user.idCheck
+            {user.idCheck
               && user.idCheck.medicareCard !== 'notAttempted'
               && this.renderItem(
                 idCheckUtils.ID_TYPE.MEDICARE_CARD,
                 user.idCheck.medicareCard,
                 user,
               )}
-            </View>
-            <View>
-              {this.renderButtons()}
-              {user.idCheck
+          </View>
+          <View>
+            {this.renderButtons()}
+            {user.idCheck
                 && (user.idCheck.australianPassport === 'matchFailed' || user.idCheck.driversLicence === 'matchFailed')
                 && (
                 <Button
@@ -256,21 +249,24 @@ class IdCheck extends PureComponent {
                   <Text>Post us certified ID</Text>
                 </Button>
                 )}
-            </View>
           </View>
-        </Content>
-      );
-    }
+        </View>
+      </Content>
+    );
+  }
 }
 
 IdCheck.propTypes = {
   user: PropTypes.object.isRequired,
+  localAuth: PropTypes.object.isRequired,
 };
 
 const mapStateToProps = (state) => {
   const user = userSelector(state);
+  const localAuth = localAuthSelector(state);
   return {
     user,
+    localAuth,
   };
 };
 
