@@ -1,14 +1,16 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import {
   View,
 } from 'react-native';
-import PropTypes from 'prop-types';
 import {
   Content,
   Button,
   Text,
 } from 'native-base';
+
+import { pinSave, localAuthSelector } from 'src/Redux/Auth';
 
 
 import {
@@ -37,8 +39,12 @@ class PinSetup extends React.Component {
 
 
     componentDidMount() {
-      const { hocs } = this.props;
+      const { hocs, localAuth } = this.props;
       const { form } = this.state;
+
+      if (localAuth.pin) {
+        // go to verification instead
+      }
 
       hocs.setForm(form);
     }
@@ -47,14 +53,18 @@ class PinSetup extends React.Component {
       const {
         screenProps,
         hocs,
+        pinSaveConnect,
+        navigation,
       } = this.props;
 
+      const next = navigation.getParam('next', '');
       const formIsValid = hocs.formIsValid();
       if (formIsValid) {
         const pin = hocs.form.pin.value;
 
         screenProps.Api.post('/user', { pin }, () => {
-          screenProps.navigateTo(routeNames.BIOMETRICS_SETUP);
+          pinSaveConnect(true);
+          screenProps.navigateTo(routeNames.BIOMETRICS_SETUP, { next });
         }, () => {
           screenProps.tastDanger('Error. Try again.');
         });
@@ -101,8 +111,24 @@ class PinSetup extends React.Component {
     }
 }
 
+PinSetup.propTypes = {
+  pinSaveConnect: PropTypes.func.isRequired,
+  localAuth: PropTypes.object.isRequired,
+};
+
+const mapStateToProps = (state) => {
+  const localAuth = localAuthSelector(state);
+  return {
+    localAuth,
+  };
+};
+
+const mapDispatchToProps = {
+  pinSaveConnect: pinSave,
+};
+
 const res = composeHoc([
   hocNames.FORM,
 ])(PinSetup);
 
-export default connect()(res);
+export default connect(mapStateToProps, mapDispatchToProps)(res);
