@@ -48,13 +48,11 @@ class FinalConfirmation extends React.Component {
             iconName: 'check-circle',
           });
           if (!user.fsMember) {
-            screenProps.navigateTo(routeNames.JOIN_FUTURE, { showBackButton:false });
+            screenProps.navigateTo(routeNames.JOIN_FUTURE, { showBackButton: false });
+          } else if (res.idCheckComplete) {
+            screenProps.navigateTo(routeNames.TAB_HOME);
           } else {
-            if (res.idCheckComplete) {
-              screenProps.navigateTo(routeNames.TAB_HOME);
-            } else {
-              screenProps.navigateTo(routeNames.ID_CHECK, { showBackButton:false });
-            }
+            screenProps.navigateTo(routeNames.ID_CHECK, { showBackButton: false });
           }
         });
       },
