@@ -1,13 +1,15 @@
 import React from 'react';
-import { Image, View, Text, StyleSheet, TextInput, KeyboardAvoidingView } from 'react-native';
+import {
+  Image, View, Text, StyleSheet, TextInput, KeyboardAvoidingView,
+} from 'react-native';
 import Button from '../ui/Button';
-const styleSettings = require('../../styles/styleSettings.json');
 import globalStyles from '../../styles/Style';
 
-export default class SummaryAndConfirm extends React.Component {
+const styleSettings = require('../../styles/styleSettings.json');
 
+export default class SummaryAndConfirm extends React.Component {
   handlePress() {
-    this.props.navigation.navigate("Feed");
+    this.props.navigation.navigate('Feed');
   }
 
   render() {
@@ -28,5 +30,4 @@ export default class SummaryAndConfirm extends React.Component {
       </View>
     );
   }
-
-};
+}
