@@ -98,7 +98,8 @@ class SmsCode extends Component {
         } else {
           screenProps.toast('Please enter the correct code');
         }
-      }).catch(() => {
+      }).catch((e) => {
+        console.log(e)
         screenProps.toast('Code timed out - Please try again');
         amplitude.getInstance().logEvent('Entered SMS Code - TimedOut', {});
         screenProps.navigateTo(routeNames.SIGN_UP_LOGIN);
