@@ -1,38 +1,38 @@
 // @flow
 
-import variable from "./../variables/platform";
+import variable from '../variables/platform';
 
 export default (variables /*: * */ = variable) => {
-  const platform = variables.platform;
+  const { platform } = variables;
 
   const tabHeadingTheme = {
-    flexDirection: "row",
+    flexDirection: 'row',
     backgroundColor: variables.tabDefaultBg,
     flex: 1,
-    alignItems: "center",
-    justifyContent: "center",
-    ".scrollable": {
+    alignItems: 'center',
+    justifyContent: 'center',
+    '.scrollable': {
       paddingHorizontal: 20,
-      flex: platform === "android" ? 0 : 1,
-      minWidth: platform === "android" ? undefined : 60
+      flex: platform === 'android' ? 0 : 1,
+      minWidth: platform === 'android' ? undefined : 60,
     },
-    "NativeBase.Text": {
+    'NativeBase.Text': {
       color: variables.topTabBarTextColor,
-      marginHorizontal: 7
+      marginHorizontal: 7,
     },
-    "NativeBase.Icon": {
+    'NativeBase.Icon': {
       color: variables.topTabBarTextColor,
-      fontSize: platform === "ios" ? 26 : undefined
+      fontSize: platform === 'ios' ? 26 : undefined,
     },
-    ".active": {
-      "NativeBase.Text": {
+    '.active': {
+      'NativeBase.Text': {
         color: variables.topTabBarActiveTextColor,
-        fontWeight: "600"
+        fontWeight: '600',
       },
-      "NativeBase.Icon": {
-        color: variables.topTabBarActiveTextColor
-      }
-    }
+      'NativeBase.Icon': {
+        color: variables.topTabBarActiveTextColor,
+      },
+    },
   };
 
   return tabHeadingTheme;
