@@ -1,41 +1,27 @@
-import React, { PureComponent } from 'react';
-import { PersistGate } from 'redux-persist/integration/react';
-import { AppRegistry } from 'react-native';
+import React, {Component} from 'react';
+import {PersistGate} from 'redux-persist/integration/react';
+import {AppRegistry} from 'react-native';
 import amplitude from 'amplitude-js';
-import './ReactotronConfig';
-import { Provider } from 'react-redux';
+import {Provider} from 'react-redux';
 import ServiceManager from 'src/Services';
-import { name as appName } from '../app.json';
-import { getStore, getPersistor } from './Redux/store';
+import {name as appName} from '../app.json';
+import {getStore, getPersistor} from './Redux/store';
 import AppIndex from './AppIndex';
 
-
-class Root extends PureComponent {
+class Root extends Component {
   componentDidMount() {
-    if (__DEV__) {
-      // Reactotron
-      Reactotron.connect();
-      console.tron = Reactotron;
-      Reactotron.clear();
-      // analytics
-    }
     this.initAnalytics();
   }
 
-
   initAnalytics() {
     if (__DEV__) {
-      amplitude.getInstance().init(
-        '827d292a9803561f32da202c0ad78a6e',
-        null,
-        { useNativeDeviceInfo: false },
-      );
+      amplitude.getInstance().init('827d292a9803561f32da202c0ad78a6e', null, {
+        useNativeDeviceInfo: false,
+      });
     } else {
-      amplitude.getInstance().init(
-        '64dd18a0f55a3c423e15a30e57d5ca30',
-        null,
-        { useNativeDeviceInfo: false },
-      );
+      amplitude.getInstance().init('64dd18a0f55a3c423e15a30e57d5ca30', null, {
+        useNativeDeviceInfo: false,
+      });
     }
   }
 
